IT WORKS!!!

+++ src/fdf.h
@@ -6,7 +6,7 @@
 /*   By: ldel-val <ldel-val@42madrid.com>          |  |           *           */
 /*                                                 \  '.___.;       +         */
 /*   Created: 2024/11/26 12:34:16 by ldel-val       '._  _.'   .        .     */
-/*   Updated: 2024/12/11 17:27:25 by ldel-val          ``                     */
+/*   Updated: 2024/12/13 17:11:03 by ldel-val          ``                     */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,12 @@
 # include <math.h>
 
 # include "color/color.h"
-# include "parse/maps/maps.h"
 
 typedef struct s_3d_point
 {
-	float	x;
-	float	y;
-	float	z;
+	int	x;
+	int	y;
+	int	z;
 	t_rgba	color;
 }	t_3d_point;
 
@@ -48,6 +47,7 @@ typedef struct s_2d_line
 	t_2d_point p2;
 }	t_2d_line;
 
+# include "parse/maps/maps.h"
 # include "utils/utils.h"
 # include "projection/projections.h"
